only delete file if it still exists #9

+++ src/ImportFileReader.php
@@ -28,6 +28,8 @@ abstract class ImportFileReader
 
     public function afterRead()
     {
-        unlink($this->file->getRealPath());
+        if(file_exists($this->file->getRealPath())) {
+            unlink($this->file->getRealPath());
+        }
     }
 }
